+++ src/types/api/userWallets.ts
@@ -1,0 +1,11 @@
+import {CreatedModified} from 'types/createdModified';
+import {Currency} from 'types/currencies';
+
+export interface UserWallet extends CreatedModified {
+  balance: number;
+  currency: Currency;
+  id: number;
+  is_owner: boolean;
+  rank: number;
+  total_users: number;
+}
