+++ src/types/wallets.ts
@@ -1,7 +1,6 @@
 import {CreatedModified} from 'types/createdModified';
 import {Currency} from 'types/currencies';
 import {Dict} from 'types/generic';
-import {PaginatedResponse} from 'types/pagination';
 
 export interface Wallet extends CreatedModified {
   balance: number;
@@ -12,16 +11,5 @@ export interface Wallet extends CreatedModified {
   id: number;
   owner: number;
 }
-
-export interface UserWallet extends CreatedModified {
-  balance: number;
-  currency: Currency;
-  id: number;
-  is_owner: boolean;
-  rank: number;
-  total_users: number;
-}
-
-export type UserWalletsPaginatedResponse = PaginatedResponse<UserWallet>;
 
 export type Wallets = Dict<Wallet>;
